% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aggregation.R
\name{create_pivot_list}
\alias{create_pivot_list}
\title{Create pivot list}
\usage{
create_pivot_list(pivot_data, arrange_cols = "process")
}
\arguments{
\item{pivot_data}{privot data as retrieved from function pivot_data()}

\item{arrange_cols}{columns used for arranging the data (default: "process")}
}
\value{
a list of results, where each element contains the result table for 
one lci_method
}
\description{
Create pivot list
}
\examples{

zipfile <- system.file("extdata/umberto-nxt_v7.1.0.13.503/Beispiel_Auswertung.zip", 
package = "kwb.umberto")
umberto7_csv_dir <- file.path(tempdir(), "Beispiel_Auswertung")
unzip(zipfile, exdir = umberto7_csv_dir)
umberto7_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto7_csv_dir)
umberto7_data_grouped <- kwb.umberto::group_data(umberto7_rawdata)
umberto7_data_pivot <- kwb.umberto::pivot_data(umberto7_data_grouped)
umberto7_data_pivot_list <- kwb.umberto::create_pivot_list(umberto7_data_pivot)
head(umberto7_data_pivot_list)

umberto10_csv_dir <- system.file("extdata/umberto-lca+_v10.1.0.3.146", 
package = "kwb.umberto")
umberto10_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto10_csv_dir)
umberto10_data_grouped <- kwb.umberto::group_data(umberto10_rawdata)
umberto10_data_pivot <- kwb.umberto::pivot_data(umberto10_data_grouped)
umberto10_data_pivot_list <- kwb.umberto::create_pivot_list(umberto10_data_pivot)
head(umberto10_data_pivot_list)

}
