% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aggregation.R
\name{group_data}
\alias{group_data}
\title{Group data}
\usage{
group_data(raw_data, grouping_paras = c("lci_method", "model", "process",
  "unit"), grouping_function = "sum")
}
\arguments{
\item{raw_data}{as retrieved by function import_rawdata()}

\item{grouping_paras}{cleaned column names used for grouping. 
(default: c("lci_method", "model", "process", "unit"))}

\item{grouping_function}{R function used for grouping (default: "sum")}
}
\value{
aggregated data according
}
\description{
Group data
}
\examples{
zipfile <- system.file("extdata/Beispiel_Auswertung.zip", package = "kwb.lca")
temp <- file.path(tempdir(), "Beispiel_Auswertung")
unzip(zipfile, exdir = temp)
rawdata <- kwb.lca::import_rawdata(csv_dir = temp)
data_grouped <- kwb.lca::group_data(rawdata)
head(data_grouped)
}
