% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_json_files_to_excel.R
\name{import_json_files_to_excel}
\alias{import_json_files_to_excel}
\title{Import JSON files to Excel File}
\usage{
import_json_files_to_excel(json_dir, file = NULL, open = TRUE)
}
\arguments{
\item{json_dir}{path to directory containing .json files}

\item{file}{path to Excel file to be created. Default: 
\code{"umberto-results_<json_dir_name>.xlsx"} in \code{tempdir()}}

\item{open}{logical indicating whether or not to open the created Excel file}
}
\value{
path to created Excel file
}
\description{
Import JSON files to Excel File
}
