% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rawdata.R
\name{import_rawdata}
\alias{import_rawdata}
\title{Import rawdata}
\usage{
import_rawdata(csv_dir, sep = ";", ...)
}
\arguments{
\item{csv_dir}{path to directory with .csv files}

\item{sep}{Single character used to separate fields within a record (default: 
in GERMAN format, i.e. sep = ";")}

\item{...}{further arguments passed to either readr::read_csv2 (default for sep == ';') or 
readr::read_csv (if sep == ',')}
}
\value{
data.frame with all imported raw data
}
\description{
Import rawdata
}
\examples{
umberto7_csv_dir <- system.file("extdata/umberto-nxt_v7.1.0.13.503", 
package = "kwb.umberto")
umberto7_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto7_csv_dir)
head(umberto7_rawdata)

umberto10_csv_dir <- system.file("extdata/umberto-lca+_v10.1.0.3.146", 
package = "kwb.umberto")
umberto10_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto10_csv_dir)
head(umberto10_rawdata)
}
