% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rawdata_json.R
\name{import_rawdata_json}
\alias{import_rawdata_json}
\title{Import Umberto results from .json files}
\usage{
import_rawdata_json(
  json_dir,
  old_format = TRUE,
  add_place = FALSE,
  files = NULL
)
}
\arguments{
\item{json_dir}{path to directory containing .json files. All .json files
will be read.}

\item{old_format}{if TRUE (the default) the same columns are provided that
are provided by \code{\link{import_rawdata}} that imports .csv files}

\item{add_place}{With add_place = TRUE, the "place" is contained in the 
result even if old_format = TRUE}

\item{files}{optional. If given and not \code{NULL} this is expected to be a
vector of character with the full paths to the \code{.json} files to be 
read.}
}
\value{
data frame
}
\description{
Import Umberto results from .json files
}
