% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aggregation.R
\name{pivot_data}
\alias{pivot_data}
\title{Make pivot data}
\usage{
pivot_data(rawdata_grouped, cols_to_ignore = "unit", key_col = "model",
  value_col = "quantity_sum")
}
\arguments{
\item{rawdata_grouped}{grouped rawdata as retrieved by function group_data()}

\item{cols_to_ignore}{column names not to be considered for transforming the 
data from the "list" to the "wide" format with tidyr::spread()
(default: "unit")}

\item{key_col}{column name containing the key for transforming the 
data from the "list" to the "wide" format with tidyr::spread() (default: "model")}

\item{value_col}{column name containing the values that will be used during 
transforming the data from the "list" to the "wide" format with tidyr::spread() 
(default: "quantity_sum")}
}
\value{
data.frame with the aggregated results for all different"lci_method"
}
\description{
Make pivot data
}
\examples{
zipfile <- system.file("extdata/Beispiel_Auswertung.zip", package = "kwb.lca")
temp <- file.path(tempdir(), "Beispiel_Auswertung")
unzip(zipfile, exdir = temp)
rawdata <- kwb.lca::import_rawdata(csv_dir = temp)
data_grouped <- kwb.lca::group_data(rawdata)
data_pivot <- kwb.lca::pivot_data(data_grouped)
head(data_pivot)
}
