% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot results}
\usage{
plot_results(
  grouped_data,
  x_col = "model",
  y_col = "quantity_sum",
  fill_col = "process"
)
}
\arguments{
\item{grouped_data}{data.frame with grouped rawdata as retrieved by 
function group_data()}

\item{x_col}{column name to be used for plotting on x-axis (default: "model")}

\item{y_col}{column name to be used for plotting on y-axis (default: 
"quantity_sum")}

\item{fill_col}{column name to be used for filling (default: "process")}
}
\value{
simple plots for all differn
}
\description{
Plot results
}
\examples{
zipfile <- system.file("extdata/umberto-nxt_v7.1.0.13.503/Beispiel_Auswertung.zip", 
package = "kwb.umberto")
umberto7_csv_dir <- file.path(tempdir(), "Beispiel_Auswertung")
unzip(zipfile, exdir = umberto7_csv_dir)
umberto7_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto7_csv_dir)
umberto7_data_grouped <- kwb.umberto::group_data(umberto7_rawdata)
kwb.umberto::plot_results(grouped_data = umberto7_data_grouped)

umberto10_csv_dir <- system.file("extdata/umberto-lca+_v10.1.0.3.146", 
package = "kwb.umberto")
umberto10_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto10_csv_dir)
umberto10_data_grouped <- kwb.umberto::group_data(umberto10_rawdata)
kwb.umberto::plot_results(grouped_data = umberto10_data_grouped)
}
