% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aggregation.R
\name{create_pivot_list}
\alias{create_pivot_list}
\title{Create pivot list}
\usage{
create_pivot_list(pivot_data, arrange_cols = "process")
}
\arguments{
\item{pivot_data}{privot data as retrieved from function pivot_data()}

\item{arrange_cols}{columns used for arranging the data (default: "process")}
}
\value{
a list of results, where each element contains the result table for 
one lci_method
}
\description{
Create pivot list
}
\examples{
zipfile <- system.file("extdata/Beispiel_Auswertung.zip", package = "kwb.lca")
temp <- file.path(tempdir(), "Beispiel_Auswertung")
unzip(zipfile, exdir = temp)
rawdata <- kwb.lca::import_rawdata(csv_dir = temp)
data_grouped <- kwb.lca::group_data(rawdata)
data_pivot <- kwb.lca::pivot_data(data_grouped)
data_pivot_list <- kwb.lca::create_pivot_list(data_pivot)
head(data_pivot)
}
