% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_json_files_to_excel.R
\name{import_json_files_to_excel}
\alias{import_json_files_to_excel}
\title{Import JSON files to Excel File}
\usage{
import_json_files_to_excel(
  json_dir,
  file = file.path(json_dir, "umberto-results.xlsx"),
  overwrite = FALSE,
  open = TRUE,
  expand_keys = c("process", "place", "exchange")
)
}
\arguments{
\item{json_dir}{path to directory containing .json files}

\item{file}{path to Excel file to be created. Default: 
\code{"umberto-results.xlsx"} within \code{json_dir}}

\item{overwrite}{whether or not to overwrite the Excel \code{file} if it 
exists. Default: \code{FALSE}.}

\item{open}{logical indicating whether or not to open the created Excel file}

\item{expand_keys}{If this argument is not \code{NULL} but a vector of (key)
column names, all sheets are expanded to the same number of rows. All
possible combinations of values in these key columns are then given in each
sheet even though there are no values for these key value combinations.
Default: \code{c("indicator", "process", "place", "exchange")}}
}
\value{
path to created Excel file
}
\description{
Import JSON files to Excel File
}
