% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_json_files_to_excel.R
\name{import_json_files_to_excel}
\alias{import_json_files_to_excel}
\title{Import JSON files to Excel File}
\usage{
import_json_files_to_excel(
  json_dir,
  file = file.path(json_dir, "umberto-results.xlsx"),
  overwrite = FALSE,
  open = TRUE,
  expand = TRUE
)
}
\arguments{
\item{json_dir}{path to directory containing .json files}

\item{file}{path to Excel file to be created. Default: 
\code{"umberto-results.xlsx"} within \code{json_dir}}

\item{overwrite}{whether or not to overwrite the Excel \code{file} if it 
exists. Default: \code{FALSE}.}

\item{open}{logical indicating whether or not to open the created Excel file}

\item{expand}{logical indicating whether or not to expand all sheets to the
same number of rows (so that all possible combinations of values in the
key columns are given)}
}
\value{
path to created Excel file
}
\description{
Import JSON files to Excel File
}
