% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rawdata.R
\name{import_rawdata}
\alias{import_rawdata}
\title{Import rawdata}
\usage{
import_rawdata(csv_dir)
}
\arguments{
\item{csv_dir}{path to directory with .csv files (in GERMAN format, i.e. 
sep = ";" and dec = ",") to be imported}
}
\value{
data.frame with all imported raw data
}
\description{
Import rawdata
}
\examples{
zipfile <- system.file("extdata/Beispiel_Auswertung.zip", package = "kwb.lca")
temp <- file.path(tempdir(), "Beispiel_Auswertung")
unzip(zipfile, exdir = temp)
rawdata <- kwb.lca::import_rawdata(csv_dir = temp)
head(rawdata)
}
