% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aggregation.R
\name{pivot_data}
\alias{pivot_data}
\title{Make pivot data}
\usage{
pivot_data(rawdata_grouped, cols_to_ignore = "unit", key_col = "model",
  value_col = "quantity_sum")
}
\arguments{
\item{rawdata_grouped}{grouped rawdata as retrieved by function group_data()}

\item{cols_to_ignore}{column names not to be considered for transforming the 
data from the "list" to the "wide" format with tidyr::spread()
(default: "unit")}

\item{key_col}{column name containing the key for transforming the 
data from the "list" to the "wide" format with tidyr::spread() (default: "model")}

\item{value_col}{column name containing the values that will be used during 
transforming the data from the "list" to the "wide" format with tidyr::spread() 
(default: "quantity_sum")}
}
\value{
data.frame with the aggregated results for all different"lci_method"
}
\description{
Make pivot data
}
\examples{
umberto7_csv_dir <- system.file("extdata/umberto-nxt_v7.1.0.13.503", 
package = "kwb.umberto")
umberto7_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto7_csv_dir)
umberto7_data_grouped <- kwb.umberto::group_data(umberto7_rawdata)
umberto7_data_pivot <- kwb.umberto::pivot_data(umberto7_data_grouped)
head(umberto7_data_pivot)

umberto10_csv_dir <- system.file("extdata/umberto-lca+_v10.1.0.3.146", 
package = "kwb.umberto")
umberto10_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto10_csv_dir)
umberto10_data_grouped <- kwb.umberto::group_data(umberto10_rawdata)
umberto10_data_pivot <- kwb.umberto::pivot_data(umberto10_data_grouped)
head(umberto10_data_pivot)

}
