% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot results}
\usage{
plot_results(grouped_data, x_col = "model", y_col = "quantity_sum",
  fill_col = "process")
}
\arguments{
\item{grouped_data}{data.frame with grouped rawdata as retrieved by 
function group_data()}

\item{x_col}{column name to be used for plotting on x-axis (default: "model")}

\item{y_col}{column name to be used for plotting on y-axis (default: 
"quantity_sum")}

\item{fill_col}{column name to be used for filling (default: "process")}
}
\value{
simple plots for all differn
}
\description{
Plot results
}
\examples{
zipfile <- system.file("extdata/Beispiel_Auswertung.zip", package = "kwb.umberto")
temp <- file.path(tempdir(), "Beispiel_Auswertung")
unzip(zipfile, exdir = temp)
rawdata <- kwb.umberto::import_rawdata(csv_dir = temp)
data_grouped <- kwb.umberto::group_data(rawdata)
kwb.umberto::plot_results(grouped_data = data_grouped)
}
