% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_export.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Write results to EXCEL}
\usage{
write_xlsx(data_pivot_list, path = "results.xlsx")
}
\arguments{
\item{data_pivot_list}{a list as retrieved by function create_pivot_list()}

\item{path}{relative or full path to be used for exporting the results to 
EXCEL (default: "results.xlsx")}
}
\value{
writes results in EXCEl file defined in "path", where each sheet
}
\description{
Write results to EXCEL
}
\examples{
zipfile <- system.file("extdata/Beispiel_Auswertung.zip", package = "kwb.lca")
temp <- file.path(tempdir(), "Beispiel_Auswertung")
unzip(zipfile, exdir = temp)
rawdata <- kwb.lca::import_rawdata(csv_dir = temp)
data_grouped <- kwb.lca::group_data(rawdata)
data_pivot <- kwb.lca::pivot_data(data_grouped)
data_pivot_list <- kwb.lca::create_pivot_list(data_pivot)
export_path <- file.path(getwd(), "results.xlsx")
print(sprintf("Exporting aggregated results to \%s", export_path))
write_xlsx(data_pivot_list, path = export_path)
}
